[ADD] petroil: Active only in Kanban and List Maintenance Requests
Adding a default filter to the menu "Maintenance Requests" that only
show the user the requests with the field `archive` in False, instead of
a domain in the action so the user can remove the filter in the other
views (not List nor Kanban).

In the case of the Maintenance Dashboard menu, we set the counters so
that they don't take into account the archived requests (`archive = True`),
and the actions the come from that view will now show those requests.

Also, hiding the button "Cancel" in the form view of the model
`maintenance.request`, for the requests that have "Preventive" as their
maintenance type.